fix(webhooks): emit observability signals before Stripe patch + consolidate person queries
- Move Sentry.captureMessage and logger.warn for Path 4 recovery before
  the subscriptions.update try block — DB writes are already committed at
  that point and observability must not depend on the best-effort patch
- Move findPersonById (from program-profile.ts) and findPersonByStripeCustomerId
  (from billing.ts) into person.ts — both query the Person model and belong
  alongside other person-lookup functions
